Move IndexDocument and its subclasses to their own file
